requirements.txt
streamlit>=1.30.0
python-docx>=0.8.11
docx2pdf>=0.1.8
language-tool-python>=2.7.1
sqlite-utils>=3.35.0
python-dateutil>=2.8.2
pywin32 -y
pywin32
pandas
plotly

